analytics retrieval implementation
