--- src/main/java/com/mailersend/sdk/AnalyticsResponse.java
@@ -1,15 +1,0 @@
-package com.mailersend.sdk;
-
-import com.google.gson.annotations.SerializedName;
-
-class AnalyticsResponse extends MailerSendResponse {
-
-    @SerializedName("data")
-    public AnalyticsList response;
-    
-    
-    protected void postDeserialize() {
-        
-        response.postDeserialize();
-    }
-}
