--- src/main/java/com/mailersend/sdk/AnalyticsStatistic.java
@@ -1,12 +1,0 @@
-package com.mailersend.sdk;
-
-import com.google.gson.annotations.SerializedName;
-
-public class AnalyticsStatistic {
-
-    @SerializedName("name")
-    public String name;
-    
-    @SerializedName("count")
-    public int count;
-}
