+++ src/main/java/com/mailersend/sdk/analytics/Analytics.java
@@ -1,9 +1,18 @@
-package com.mailersend.sdk;
+/*************************************************
+ * MailerSend Java SDK
+ * https://github.com/mailersend/mailersend-java
+ * 
+ * @author MailerSend <support@mailersend.com>
+ * https://mailersend.com
+ **************************************************/
+package com.mailersend.sdk.analytics;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Date;
 
+import com.mailersend.sdk.MailerSend;
+import com.mailersend.sdk.MailerSendApi;
 import com.mailersend.sdk.exceptions.MailerSendException;
 
 public class Analytics {
@@ -20,12 +29,21 @@ public class Analytics {
     private ArrayList<String> tagsFilter = new ArrayList<String>();
        
     
-    protected Analytics(MailerSend ref) {
+    /**
+     * Do not initialize directly. This should only be accessed from MailerSend.analytics
+     * @param apiReference
+     */
+    public Analytics(MailerSend ref) {
         
         apiObjectReference = ref;
     }
     
     
+    /**
+     * Sets the domain id for the request
+     * @param domainId
+     * @return
+     */
     public Analytics domainId(String domainId) {
         
         this.domainIdFilter = domainId;
@@ -33,6 +51,12 @@ public class Analytics {
         return this;
     }
     
+    
+    /**
+     * Sets the date from filter for the request
+     * @param dateFrom
+     * @return
+     */
     public Analytics dateFrom(Date dateFrom) {
         
         this.dateFromFilter = dateFrom;
@@ -40,6 +64,12 @@ public class Analytics {
         return this;
     }
     
+    
+    /**
+     * Sets the date to filter for the request
+     * @param dateTo
+     * @return
+     */
     public Analytics dateTo(Date dateTo) {
         
         this.dateToFilter = dateTo;
@@ -48,6 +78,11 @@ public class Analytics {
     }
     
     
+    /**
+     * Sets the tags filters for the request
+     * @param tags
+     * @return
+     */
     public Analytics tags(String[] tags) {
         
         this.tagsFilter.clear();
@@ -56,6 +91,12 @@ public class Analytics {
         return this;
     }
     
+    
+    /**
+     * Adds a tag filter to the request
+     * @param tag
+     * @return
+     */
     public Analytics tag(String tag) {
         
         this.tagsFilter.add(tag);
@@ -64,13 +105,36 @@ public class Analytics {
     }
     
     
-    public AnalyticsByDateList getByDate() throws MailerSendException {
+    /**
+     * Gets the analytics by date
+     * @param events
+     * @return
+     * @throws MailerSendException
+     */
+    public AnalyticsByDateList getByDate(String[] events) throws MailerSendException {
         
-        return this.getByDate("days", null);
+        return this.getByDate("group_by=days", events);
     }
     
     
+    /**
+     * Gets the analytics by date
+     * @param groupBy
+     * @param events
+     * @return
+     * @throws MailerSendException
+     */
     public AnalyticsByDateList getByDate(String groupBy, String[] events) throws MailerSendException {
+        
+        if (dateFromFilter == null || dateToFilter == null) {
+            
+            throw new MailerSendException("Date from and Date to dates are required.");
+        }
+        
+        if (events == null || events.length == 0) {
+            
+            throw new MailerSendException("No events passed.");
+        }
         
         String endpoint = "/analytics/date";
         
@@ -86,7 +150,7 @@ public class Analytics {
         endpoint = endpoint.concat(prepareParamsUrl(params.toArray(new String[0])));
         
         MailerSendApi api = new MailerSendApi();
-        api.setToken(apiObjectReference.token);
+        api.setToken(apiObjectReference.getToken());
         
         AnalyticsByDateResponse msResponse = api.getRequest(endpoint, AnalyticsByDateResponse.class);
         
@@ -96,30 +160,51 @@ public class Analytics {
     }
     
     
+    /**
+     * Gets the opens by country
+     * @return
+     * @throws MailerSendException
+     */
     public AnalyticsList getOpensByCountry() throws MailerSendException {
         
-        return genericAnalyticsRequest("analytics/country");
+        return genericAnalyticsRequest("/analytics/country");
     }
     
     
+    /**
+     * Gets the opens by user agent
+     * @return
+     * @throws MailerSendException
+     */
     public AnalyticsList getOpensByUserAgent() throws MailerSendException {
         
-        return genericAnalyticsRequest("analytics/ua-name");
+        return genericAnalyticsRequest("/analytics/ua-name");
     }
     
     
+    /**
+     * Gets the opens by the user agent type (by reading environment)
+     * @return
+     * @throws MailerSendException
+     */
     public AnalyticsList getOpensByUserAgenType() throws MailerSendException {
         
-        return genericAnalyticsRequest("analytics/ua-type");
+        return genericAnalyticsRequest("/analytics/ua-type");
     }
     
     
+    /**
+     * Implementation for all analytics requests except the getByDate ones
+     * @param requestEndpoint
+     * @return
+     * @throws MailerSendException
+     */
     private AnalyticsList genericAnalyticsRequest(String requestEndpoint) throws MailerSendException {
         
-        String endpoint = requestEndpoint.concat(arrayToUrlRequest(null, null));
+        String endpoint = requestEndpoint.concat(prepareParamsUrl(null));
         
         MailerSendApi api = new MailerSendApi();
-        api.setToken(apiObjectReference.token);
+        api.setToken(apiObjectReference.getToken());
         
         AnalyticsResponse msResponse = api.getRequest(endpoint, AnalyticsResponse.class);
         
@@ -129,12 +214,20 @@ public class Analytics {
     }
     
     
+    /**
+     * Prepares the query part of the request url
+     * @param additionalParams pass additional params in the form of name=value to include to the request url
+     * @return
+     */
     private String prepareParamsUrl(String[] additionalParams) {
         
         // prepare the request parameters
         ArrayList<String> params = new ArrayList<String>();
         
-        params.addAll(Arrays.asList(additionalParams));
+        if (additionalParams != null) {
+        
+            params.addAll(Arrays.asList(additionalParams));
+        }
         
         if (domainIdFilter != null) {
             
@@ -173,7 +266,18 @@ public class Analytics {
     }
     
     
+    /**
+     * Builds a GET array query for the values with name field. E.g. field[]=value1&field[]=value2
+     * @param values
+     * @param field
+     * @return
+     */
     private String arrayToUrlRequest(String[] values, String field) {
+        
+        if (values == null || field == null) {
+            
+            return "";
+        }
         
         String params = "";
