+++ src/main/java/com/mailersend/sdk/analytics/AnalyticsByDateResponse.java
@@ -1,0 +1,26 @@
+/*************************************************
+ * MailerSend Java SDK
+ * https://github.com/mailersend/mailersend-java
+ * 
+ * @author MailerSend <support@mailersend.com>
+ * https://mailersend.com
+ **************************************************/
+package com.mailersend.sdk.analytics;
+
+import com.google.gson.annotations.SerializedName;
+import com.mailersend.sdk.MailerSendResponse;
+
+public class AnalyticsByDateResponse extends MailerSendResponse {
+
+    @SerializedName("data")
+    public AnalyticsByDateList response;
+    
+    
+    /**
+     * Does all the needed actions after deserialization 
+     */
+    public void postDeserialize() {
+        
+        response.postDeserialize();
+    }
+}
